- Fixed many compiler warnings that were due to synthesizing atomic properties. Made them all nonatomic. - Fixed lots of deprecated method warnings. - Fixed a bunch of signed/unsigned warnings. - Fixed issue where editing image and image borders were misaligned. - Made a few things iPad only. - Fixed issue where picture gallery was using fullscreen layout, causing pictures to sometimes be scrolled strangely.

+++ klodder/Classes/UndoStack.cpp
@@ -63,14 +63,14 @@ void UndoState::SetDataLayerImage(int imageIndex, int index, const MacImage* ima
 	m.mImageLocation.Set(x, y);
 }
 
-void UndoState::SetCommandStreamLocation(size_t location)
+void UndoState::SetCommandStreamLocation(int location)
 {
 	Assert(!mHasCommandStreamLocation);
 	mHasCommandStreamLocation = true;
 	mCommandStreamLocation = location;
 }
 
-void UndoState::SetDataStreamLocation(size_t location)
+void UndoState::SetDataStreamLocation(int location)
 {
 	Assert(!mHasDataStreamLocation);
 	mHasDataStreamLocation = true;
